+++ packages/gallery/src/versionLogger.js
@@ -1,1 +1,1 @@
-try {console.debug('Pro Gallery Version: 2.1.0'); } catch (e) {}
+try {console.debug('Pro Gallery Version: 2.1.2'); } catch (e) {}