+++ packages/lib/src/core/blueprints/Blueprints.js
@@ -38,7 +38,7 @@ class Blueprints {
         existingBlueprint.structure = structure;
 
         // if its an infinite gallery - let the container loose
-        const isInfinite = existingBlueprint.styles.oneRow && existingBlueprint.styles.enableInfiniteScroll
+        const isInfinite = !existingBlueprint.styles.oneRow && existingBlueprint.styles.enableInfiniteScroll
         if (isInfinite) {
           existingBlueprint.container.height = structure.height;
         }