+++ packages/lib/src/core/presets/panoramaGallery.js
@@ -1,5 +1,6 @@
 import LAYOUTS from '../../common/constants/layout';
 import SCROLL_DIRECTION from '../../common/constants/scrollDirection';
+import {calcTargetItemSize} from '../helpers/layoutHelper';
 
 export const fixedStyles = {
   galleryLayout: LAYOUTS.PANORAMA,
@@ -27,6 +28,6 @@ export const createStyles = styles => {
   return {
     ...styles,
     ...fixedStyles,
-    targetItemSize: styles.targetItemSize || styles.gallerySize,
+    targetItemSize: calcTargetItemSize(styles),
   }
 }